+++ Core/Subscale/DenseUnitGenerator/DenseUnit.h
@@ -9,13 +9,13 @@ class DenseUnit {
 private:
 	Points points_;
 	uint64_t signature_;
-	uint64_t dimension_;
+    uint32_t dimension_;
 
 public:
-	DenseUnit(Points points, uint64_t signature, uint64_t dimension);
+	DenseUnit(Points points, uint64_t signature, uint32_t dimension);
 	Points getPoints();
 	uint64_t getSignature() const;
-    uint64_t getDimension() const;
+    uint32_t getDimension() const;
 };
 
 typedef std::vector<DenseUnit> DenseUnits;
